pyproject.toml
[tool.poetry]
name = "weaviate-hybrid"
version = "0.1.0"
description = ""
authors = ["Robert Alexander <gogonegro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
weaviate-client = "^4.4.4"
python-dotenv = "^1.0.1"
stanza = "^1.7.0"
pandas = "^2.2.0"
matplotlib = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
